@@ -7,11 +7,23 @@ import { AppRoutingModule } from './app-routing.module'
 import { AppComponent } from './app.component'
 import { SharedModule } from './feature-modules/shared/shared.module'
 import { LoginComponent } from './components/login/login.component'
-import { HttpClientModule } from '@angular/common/http';
+import { HttpClientModule } from '@angular/common/http'
 import { HomeComponent } from './components/home/home.component'
+import { SectionComponent } from './components/section/section.component'
+import { PostCreateComponent } from './components/post-create/post-create.component'
+import { NavbarComponent } from './components/navbar/navbar.component'
+import { FreeInputDirective } from './directives/free-input.directive'
 
 @NgModule({
-  declarations: [AppComponent, LoginComponent, HomeComponent],
+  declarations: [
+    AppComponent,
+    LoginComponent,
+    HomeComponent,
+    SectionComponent,
+    PostCreateComponent,
+    NavbarComponent,
+    FreeInputDirective,
+  ],
   imports: [
     BrowserModule,
     BrowserAnimationsModule,
