@@ -1,10 +1,15 @@
-import { Component } from '@angular/core';
+import { Component } from '@angular/core'
+import { AuthService } from './services/auth/auth.service'
 
 @Component({
   selector: 'app-root',
   templateUrl: './app.component.html',
-  styleUrls: ['./app.component.scss']
+  styleUrls: ['./app.component.scss'],
 })
 export class AppComponent {
-  title = 'napredne-baze-elektrijada';
+  constructor(private authService: AuthService) {}
+
+  get isAuthenticated(): boolean {
+    return this.authService.isAuthenticated
+  }
 }
